bacon: Revert to old RIL and QMI framework
New RIL suffers from frequent, spontaneous data disconnects. Data connection
is reported to be fine, and signal is strong, yet data is not connected. The
connection does not automatically restore itself either because the RIL reports
everything to be good.

Revert to the old RIL in order to fix the data drop-outs.