onyx: Probe rear camera first on init
Sometimes at boot, the front-facing camera is assigned camera ID 0 because
it is registered before the rear camera. This leads to all sorts of mayhem,
ultimately causing the rear camera to be missing after boot in not-so-rare
cases, and not reappear until a reboot.

Hex-edit the sensor modules blob in order to always probe the rear camera
sensor first, so that the cameras are registered in the order that Android
expects.